refactor(analysis-types): extract baseline metrics DTO
